Implement allocation alignment
The only supported CL implementation is currently SBCL. On all others,
MAKE-STATIC-VECTOR will signal an error if the keyword arg ALIGNMENT
is passed.

The implementation uses posix_memalign() to return aligned memory, and
it adds another header before SBCL's array header, to keep track of
the pointer originally returned by the allocator.
Because the two headers must be placed before the data, which is
aligned, it will waste some amount of memory.